requirements.txt
scapy
google-generativeai
python-dotenv
urllib3
pathlib